+++ feature/history/src/main/java/dev/diegoflassa/bipsale/feature/history/SaleDetailScreen.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import dev.diegoflassa.bipsale.core.domain.model.SaleItem
 import dev.diegoflassa.bipsale.core.ui.theme.BipSaleTheme
 import java.text.SimpleDateFormat